OK letterbox.c and merge it with lifebar.c
