+++ versions/us1/mischiefmakers.yaml
@@ -98,7 +98,6 @@ segments:
       - [0x01E7A0, c, actor_init]
       - [0x01F1E0, c]
       - [0x022290, c, letterbox]
-      - [0x023070, c, lifebar]
       - [0x023910, c, soft_reset]
       - [0x0241E0, c] # has to do with scenes (scene init?)
       - [0x026A00, c] # might be a fake split
@@ -317,7 +316,7 @@ segments:
       - [0x0C8DE0, .data, stage]
       - [0x0CA300, .data, actor_init]
       - [0x0CADC0, .data, 1F1E0]
-      - [0x0CCA20, .data, lifebar]
+      - [0x0CCA20, .data, letterbox]
       - [0x0CCB40, .data, soft_reset]
       - [0x0CCB50, .data, 241E0]
       - [0x0CDC80, .data, 26A00]
